chore: update comments and moved reaction types
