+++ packages/core/src/React/hooks/use-respond-to-reaction-response/types.ts
@@ -1,5 +1,7 @@
-import { TrixtaReactionInstance } from '../../types/common';
-import { TrixtaReactionBaseProps } from '../../types/reactions';
+import {
+  TrixtaReactionBaseProps,
+  TrixtaReactionInstance,
+} from '../../types/reactions';
 
 export interface UseRespondToReactionResponseProps
   extends TrixtaReactionBaseProps {
