+++ packages/core/src/React/hooks/use-respond-to-reaction-response/use-respond-to-reaction-response.ts
@@ -7,12 +7,11 @@ import {
   makeSelectTrixtaReactionResponseInstancesForRole,
 } from '../../selectors';
 import { trixtaDebugger, TrixtaDebugType } from '../../TrixtaDebugger';
+import { DefaultUnknownType, TrixtaState } from '../../types/common';
 import {
-  DefaultUnknownType,
+  TrixtaReactionBaseProps,
   TrixtaReactionInstance,
-  TrixtaState,
-} from '../../types/common';
-import { TrixtaReactionBaseProps } from '../../types/reactions';
+} from '../../types/reactions';
 import {
   RespondToReactionFunctionParameters,
   UseRespondToReactionResponseHookReturn,
